Generating a lot of interesting line drawings rn
This version is working and has taken to generating a lot of line diagrams right now. Visually interesting maybe but not where I want to take this. Good commit to come back to later on if I want to fork this and add some more interesting noise/random/perlin elements.